+++ packages/vue/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const webpackConfigPath = require.resolve(
-    '@vue/cli-service/webpack.config.js',
-);
+// - Corrige la prise en charge des plugins chargés dans les configs. partagées.
+// @see https://github.com/eslint/eslint/issues/3458
+require('@rushstack/eslint-patch/modern-module-resolution');
 
 module.exports = {
     // - Parseur
@@ -12,6 +12,9 @@ module.exports = {
         sourceType: 'module',
         ecmaFeatures: {
             jsx: true,
+        },
+        vueFeatures: {
+            interpolationAsNonHTML: false,
         },
     },
 
@@ -30,6 +33,7 @@ module.exports = {
             node: {
                 extensions: [
                     '.mjs',
+                    '.cjs',
                     '.js',
                     '.jsx',
                     '.ts',
@@ -38,7 +42,6 @@ module.exports = {
                     '.json',
                 ],
             },
-            webpack: { config: webpackConfigPath },
         },
         'react': {
             pragma: 'h',
@@ -65,7 +68,7 @@ module.exports = {
     rules: {
         // - L'extension ne doit pas être spécifiée dans les imports de fichiers contenant du JavaScript.
         //   (Le support des fichiers `.vue` est ajouté dans cet overwrite)
-        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/extensions.md
+        // @see https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/extensions.md
         'import/extensions': ['error', 'ignorePackages', {
             vue: 'never',
             js: 'never',
@@ -79,9 +82,10 @@ module.exports = {
         // @see https://eslint.vuejs.org/rules/html-indent.html
         'vue/html-indent': ['error', 4],
 
-        // enforce that class methods use "this"
+        // - Les méthodes de classe doivent utiliser `this` ou être extraites dans des fonctions pures externes à la classe.
         // @see https://eslint.org/docs/rules/class-methods-use-this
         'class-methods-use-this': ['error', {
+            enforceForClassFields: true,
             exceptMethods: [],
         }],
 
@@ -133,9 +137,25 @@ module.exports = {
         //
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-pascal-case.md
         'react/jsx-pascal-case': ['error', {
+            allowLeadingUnderscore: true,
             allowAllCaps: false,
             ignore: [],
         }],
+
+        // - Empêche l'utilisation de valeurs invalides pour certains attributs HTML :
+        //   - Vérifie que la valeur des attributs `rel` est correcte en fonction de la balise l'utilisant.
+        // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-invalid-html-attribute.md
+        'react/no-invalid-html-attribute': ['error', ['rel']],
+
+        // - Cf. la règle `array-bracket-spacing` (cette version ne fait que prendre en charge la règle dans les `<template>`).
+        // @see https://eslint.org/docs/rules/arrow-spacing
+        // @see https://eslint.vuejs.org/rules/array-bracket-spacing.html
+        'vue/array-bracket-spacing': ['error', 'never'],
+
+        // Cf. la règle `arrow-spacing` (cette version ne fait que prendre en charge la règle dans les `<template>`).
+        // @see https://eslint.org/docs/rules/arrow-spacing
+        // @see https://eslint.vuejs.org/rules/arrow-spacing.html
+        'vue/arrow-spacing': ['error', { before: true, after: true }],
 
         // - S'assure de la non utilisation des tirets pour les props de components.
         // @see https://eslint.vuejs.org/rules/attribute-hyphenation.html
@@ -143,10 +163,39 @@ module.exports = {
             ignore: [],
         }],
 
+        // Cf. la règle `block-spacing` (cette version ne fait que prendre en charge la règle dans les `<template>`).
+        // @see https://eslint.org/docs/rules/block-spacing
+        // @see https://eslint.vuejs.org/rules/block-spacing
+        'vue/block-spacing': ['error', 'always'],
+
+        // - Contrôle la gestion des nouvelles lignes dans la déclation des blocks Vue (`<template>`, `<script>`) :
+        //   - Il doit toujours y avoir un saut de ligne avant et après le contenu des blocks.
+        //   - Il ne doit pas y avoir de lignes vides au début et à la fin des blocks.
+        // @see https://eslint.vuejs.org/rules/block-tag-newline
+        'vue/block-tag-newline': ['error', {
+            singleline: 'always',
+            multiline: 'always',
+            maxEmptyLines: 0,
+        }],
+
+        // Cf. la règle `brace-style` (cette version ne fait que prendre en charge la règle dans les `<template>`).
+        // @see https://eslint.org/docs/rules/brace-style
+        // @see https://eslint.vuejs.org/rules/brace-style
+        'vue/brace-style': ['error', '1tbs', { allowSingleLine: true }],
+
         // - Ordre des éléments dans les fichiers vue: <template>, <style> puis <script>.
         // @see https://eslint.vuejs.org/rules/component-tags-order.html
         'vue/component-tags-order': ['error', {
             order: ['template', 'style', 'script'],
+        }],
+
+        // - Contrôle le positionnement du premier attribut dans les composants :
+        //   - Si le composant est sur une ligne :
+        //   - Si le composant est sur plusieurs lignes :
+        // @see https://eslint.vuejs.org/rules/first-attribute-linebreak.html
+        'vue/first-attribute-linebreak': ['error', {
+            singleline: 'beside',
+            multiline: 'below',
         }],
 
         // - Utilise les auto-fermeture pour tous les types de tags.
@@ -162,17 +211,60 @@ module.exports = {
             },
         }],
 
-        // - Contrôle le nombre de props par ligne:
+        // - Contrôle le nombre de props par ligne :
         //   - Pas de limite pour les components single-line.
         //   - Une seule prop. par ligne sur les components multi-ligne.
         // @see https://eslint.vuejs.org/rules/max-attributes-per-line.html
         'vue/max-attributes-per-line': ['error', {
-            singleline: { max: Infinity, allowFirstLine: true },
-            multiline: { max: 1, allowFirstLine: false },
+            singleline: { max: Infinity },
+            multiline: { max: 1 },
+        }],
+
+        // - Interdit l'usage de l'attribut `tag` (déprécié en Vue 3.x) dans les `<RouterLink>` (ou `<router-link>`).
+        // @see https://eslint.vuejs.org/rules/no-deprecated-router-link-tag-prop.html
+        'vue/no-deprecated-router-link-tag-prop': ['error'],
+
+        // - Interdit l'utilisation de `expose({ ... })` (Vue 3) après un `await` => Doit obligatoirement être synchrone.
+        // @see https://eslint.vuejs.org/rules/no-expose-after-await.html
+        'vue/no-expose-after-await': ['error'],
+
+        // - Interdit l'utilisation de `v-text` au profit des `children`.
+        // @see https://eslint.vuejs.org/rules/no-v-text.html
+        'vue/no-v-text': ['error'],
+
+        // - Interdit l'utilisation de propriétés non définies dans les components.
+        // @see https://eslint.vuejs.org/rules/no-undef-properties.html
+        'vue/no-undef-properties': ['error', {
+            'ignores': ['/^\\$/'],
+        }],
+
+        // Cf. la règle `no-useless-concat` (cette version ne fait que prendre en charge la règle dans les `<template>`).
+        // @see https://eslint.org/docs/rules/no-useless-concat
+        // @see https://eslint.vuejs.org/rules/no-useless-concat
+        'vue/no-useless-concat': ['error'],
+
+        // - Empêche l'utilsation de chaînes quotées inutilement (e.g. `{{ 'Hello' }}`).
+        // @see https://eslint.vuejs.org/rules/no-useless-mustaches
+        'vue/no-useless-mustaches': ['error', {
+            ignoreIncludesComment: false,
+            ignoreStringEscape: false,
         }],
 
         //
-        // - Disabled rules.
+        // - Règles désactivées (spécifiques à Vue 3).
+        //
+
+        // @see https://eslint.vuejs.org/rules/valid-define-emits
+        'vue/valid-define-emits': ['off'],
+
+        // @see https://eslint.vuejs.org/rules/valid-define-props
+        'vue/valid-define-props': ['off'],
+
+        // @see https://eslint.vuejs.org/rules/valid-v-memo
+        'vue/valid-v-memo': ['off'],
+
+        //
+        // - Règles désactivées.
         //
 
         // @see https://eslint.org/docs/rules/no-underscore-dangle
@@ -226,6 +318,9 @@ module.exports = {
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-array-index-key.md
         'react/no-array-index-key': ['off'],
 
+        // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-arrow-function-lifecycle.md
+        'react/no-arrow-function-lifecycle': ['off'],
+
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-danger.md
         'react/no-danger': ['off'],
 
@@ -250,6 +345,9 @@ module.exports = {
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-is-mounted.md
         'react/no-is-mounted': ['off'],
 
+        // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-namespace.md
+        'react/no-namespace': ['off'],
+
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-redundant-should-component-update.md
         'react/no-redundant-should-component-update': ['off'],
 
@@ -270,6 +368,9 @@ module.exports = {
 
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-unused-prop-types.md
         'react/no-unused-prop-types': ['off'],
+
+        // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/prefer-exact-props.md
+        'react/prefer-exact-props': ['off'],
 
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/state-in-constructor.md
         'react/state-in-constructor': ['off'],
@@ -298,8 +399,25 @@ module.exports = {
         // @see https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/sort-comp.md
         'react/sort-comp': ['off'],
 
+        // @see https://eslint.vuejs.org/rules/array-bracket-newline.html
+        'vue/array-bracket-newline': ['off'],
+
         // @see https://eslint.vuejs.org/rules/attributes-order.html
         'vue/attributes-order': ['off'],
+
+        // - Désactivé car nous utilisons `babel-preset-vca-jsx` qui nous permet de définir les
+        //   fonctions `setup` de manière simplifiée avec un sucre syntaxique.
+        // @see https://eslint.vuejs.org/rules/component-api-style.html
+        'vue/component-api-style': ['off'],
+
+        // @see https://eslint.vuejs.org/rules/no-child-content.html
+        'vue/no-child-content': ['off'],
+
+        // @see https://eslint.vuejs.org/rules/multi-word-component-names.html
+        'vue/multi-word-component-names': ['off'],
+
+        // @see https://eslint.vuejs.org/rules/no-restricted-class.html
+        'vue/no-restricted-class': ['off'],
 
         // @see https://eslint.vuejs.org/rules/singleline-html-element-content-newline.html
         'vue/singleline-html-element-content-newline': ['off'],
@@ -307,6 +425,6 @@ module.exports = {
 
     // - Overrides
     overrides: [
-        { files: ['**/*.ts?(x)'], ...require('./overrides/typescript') },
+        { files: ['**/*.ts?(x)'], ...require('./overrides/typescript.js') },
     ],
 };